Stock table functionality
Created functionality for the Stock table.

New Class Stock: Object representation of a stock entry in the stock table. It contains the same fields as the stock table from the database. It also contains getters, setters, and a get data method.

Four new methods in the Database class.

createGenreLink: Linkes a media entry and genre entry through the use of the genre_linker table.

createStock: Creates a stock entry connected to a media entry.

modifyStock: Modifies a stock entry.

deleteStock: Deletes a stock entry.