createBandLink method
Created a new method in Database as well as modified Media.

createBandLink: Links a media object with a band through the band ID.

Removed bandID field from Media constructor. This allows for a media object to be created without the need of a band (covering some fringe cases of music without a band and preventing other future issues).