Requests will now be displayed on Request page
Still very buggy and lacks full functionality. For example, the edit and delete buttons don't work yet on the cards. Also, the vBox isn't scrollable so once a certain number of requests is reached the vBox will start growing and cover other parts of the program. That will be fixed though. Overall though, you are able to visualize the addition and collection of requests to and from the database.